+++ tests/parser.cpp
@@ -741,21 +741,24 @@ TEST(Parser, if_elseif_elseif_else)
 
 TEST(Parser, unroll)
 {
-  test("kprobe:sys_open { $i = 0; unroll(5) { printf(\"i: %d\\n\", $i); $i = $i + 1; } }",
+  test("kprobe:sys_open { $i = 0; unroll(5) { printf(\"i: %d\\n\", $i); $i = "
+       "$i + 1; } }",
        "Program\n"
        " kprobe:sys_open\n"
        "  =\n"
        "   variable: $i\n"
        "   int: 0\n"
-       "  unroll 5\n"
-       "   call: printf\n"
-       "    string: i: %d\\n\n"
-       "    variable: $i\n"
-       "   =\n"
-       "    variable: $i\n"
-       "    +\n"
+       "  unroll\n"
+       "   int: 5\n"
+       "   block\n"
+       "    call: printf\n"
+       "     string: i: %d\\n\n"
        "     variable: $i\n"
-       "     int: 1\n");
+       "    =\n"
+       "     variable: $i\n"
+       "     +\n"
+       "      variable: $i\n"
+       "      int: 1\n");
 }
 
 TEST(Parser, ternary_str)